Randomly place cards and handleClick
The handleClick function is tied to the scoreboard using props.
Theoretically, the way this is written should mean that the contents
of the cards should be able to be replaced with something else and the
app would still function properly (given necessary changes when made
in the card component to reflect the changing data).