Refactor Cards to use a randomizedCards state
This seperates the 'cards' state variable from the 'randomizedCards'
state variable and allows ids to be assigned to cards based on their
index in the 'cards' array which does not change, as opposed to
their index in the 'randomizedCards' which changes on every click. This
means we do not need to specifically assign ids to every element, but
can instead easily assign ids based on an unchanging index.